Package structure refactoring. Update User class to store raw response data.
