+++ anaplan_api/src/FileDownload.py
@@ -6,7 +6,7 @@ from anaplan_api.src.models.AnaplanVersion import AnaplanVersion
 
 if TYPE_CHECKING:
     from anaplan_api.src.models.AnaplanConnection import AnaplanConnection
-    from .AnaplanResourceFile import AnaplanResourceFile
+    from anaplan_api.src.models.AnaplanResourceFile import AnaplanResourceFile
     from .util.RequestHandler import RequestHandler
 
 logger = logging.getLogger(__name__)
