+++ anaplan_api/src/authentication/CertificateAuthentication.py
