+++ anaplan_api/src/bases/AnaplanResource.py
